Blockmaskgenerator

+++ utils/label_selection.py
@@ -14,8 +14,9 @@ logger = logging.getLogger('Cityscapes adaptation')
 
 def kc_parameters(conf_dict, pred_cls_num, tgt_portion, round_idx, args=None):
     """args"""
-    num_classes = 19
+    num_classes = args.num_classes
 
+    # import pdb; pdb.set_trace()
 
     print('###### Start kc generation in round {} ! ######'.format(round_idx))
     start_kc = time.time()
@@ -25,7 +26,8 @@ def kc_parameters(conf_dict, pred_cls_num, tgt_portion, round_idx, args=None):
     cls_size = np.zeros(num_classes, dtype=np.float32)
 
     for idx_cls in np.arange(0, num_classes):
-        cls_size[idx_cls] = pred_cls_num[idx_cls]
+        cls_size[idx_cls] = pred_cls_num[idx_cls] # 클래스 사이즈에 pred_cls_num 의 값을 갖고 온다.
+
         if conf_dict[idx_cls] != None:
             conf_dict[idx_cls].sort(reverse=True)  # sort in descending order
             len_cls = len(conf_dict[idx_cls])
@@ -33,7 +35,7 @@ def kc_parameters(conf_dict, pred_cls_num, tgt_portion, round_idx, args=None):
             len_cls_thresh = int(cls_sel_size[idx_cls])
             if len_cls_thresh != 0:
                 cls_thresh[idx_cls] = conf_dict[idx_cls][len_cls_thresh-1]
-            conf_dict[idx_cls] = None
+            # conf_dict[idx_cls] = None
     
     print("Per class thresholds:")
     print(cls_thresh)
@@ -44,9 +46,6 @@ def kc_parameters(conf_dict, pred_cls_num, tgt_portion, round_idx, args=None):
 
 
 def label_selection(cls_thresh, round_idx, save_prob_path, save_pred_path, save_pseudo_label_path, save_pseudo_label_color_path, save_round_eval_path, args= None):
-    """args"""
-    debug = True
-
 
     print('###### Start pseudo-label generation in round {} ! ######'.format(round_idx))
     start_pl = time.time()
@@ -64,20 +63,20 @@ def label_selection(cls_thresh, round_idx, save_prob_path, save_pred_path, save_
         weighted_prob = pred_prob / cls_thresh
         weighted_prob_ids = weighted_prob.argmax(axis=2).astype(np.uint8)
 
-        if debug:
+        if len(os.listdir(save_wpred_vis_path)) < 10:
             colorize_mask(weighted_prob_ids).save('%s/%s_color.png' % (save_wpred_vis_path, sample_name))
 
         weighted_conf = weighted_prob.max(axis=2)
         pred_label_labelIDs = weighted_prob_ids
         pred_label_labelIDs[weighted_conf < 1] = 255  # '255' in cityscapes indicates 'unlabaled' for trainIDs
 
-        if debug:
+        if len(os.listdir(save_pseudo_label_color_path)) < 10:
             pseudo_label_col = colorize_mask(pred_label_labelIDs)
             pseudo_label_col.save('%s/%s_color.png' % (save_pseudo_label_color_path, sample_name))
         
         pseudo_label_save = Image.fromarray(pred_label_labelIDs.astype(np.uint8))
         pseudo_label_save.save('%s/%s.png' % (save_pseudo_label_path, sample_name))
 
-    
-    shutil.rmtree(save_prob_path)
+    if args.alpha > 0:
+        shutil.rmtree(save_prob_path)
     print('###### Finish pseudo-label generation in round {}! Time cost: {:.2f} seconds. ######'.format(round_idx, time.time() - start_pl))
